Refactor packages
Make more demo friendly
