+++ post-redirect-get/src/main/java/eu/agilejava/mvc/prg/Messages.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package eu.agilejava.mvc.domain;
+package eu.agilejava.mvc.prg;
 
 import java.io.Serializable;
 import java.util.ArrayList;
